refactor(electron): split ProcessingView into wt-* subcomponents [WTEL-9965](https://webitel.atlassian.net/browse/WTEL-9965)
- move ProcessingView, ProgressTimer, SchedulePicker into ProcessingView/
  with per-component dirs matching CallView pattern
- extract ProcessingReport (success question, wt-button) and
  ProcessingSchedule (wt-radio choice + SchedulePicker + wt-button save)
- container stays thin: useAutoResize + IPC wiring via semantic events
- drop DOM ids, inline styles, hand-rolled radio/button CSS
